[cache-machine] minor fix
